src/VumbaSoft.AdventureWorks.Controllers/Demografic/ContinentRegions/ContinentRegionsController.cs
using Microsoft.AspNetCore.Mvc;
using VumbaSoft.AdventureWorks.Objects;
using VumbaSoft.AdventureWorks.Services;
using VumbaSoft.AdventureWorks.Validators;
using System;

namespace VumbaSoft.AdventureWorks.Controllers.Demografic
{
    [Area("Demografic")]
    public class ContinentRegionsController : ValidatedController<IContinentRegionValidator, IContinentRegionService>
    {
        public ContinentRegionsController(IContinentRegionValidator validator, IContinentRegionService service)
            : base(validator, service)
        {
        }

        [HttpGet]
        public ViewResult Index()
        {
            return View(Service.GetViews());
        }

        [HttpGet]
        public ViewResult Create()
        {
            return View();
        }

        [HttpPost]
        public ActionResult Create(ContinentRegionView region)
        {
            if (!Validator.CanCreate(region))
                return View(region);

            Service.Create(region);

            return RedirectToAction("Index");
        }

        [HttpGet]
        public ActionResult Details(Int32 id)
        {
            return NotEmptyView(Service.Get<ContinentRegionView>(id));
        }

        [HttpGet]
        public ActionResult Edit(Int32 id)
        {
            return NotEmptyView(Service.Get<ContinentRegionView>(id));
        }

        [HttpPost]
        public ActionResult Edit(ContinentRegionView region)
        {
            if (!Validator.CanEdit(region))
                return View(region);

            Service.Edit(region);

            return RedirectToAction("Index");
        }

        [HttpGet]
        public ActionResult Delete(Int32 id)
        {
            return NotEmptyView(Service.Get<ContinentRegionView>(id));
        }

        [HttpPost]
        [ActionName("Delete")]
        public RedirectToActionResult DeleteConfirmed(Int32 id)
        {
            Service.Delete(id);

            return RedirectToAction("Index");
        }
    }
}
